fix:removed opencv dependencies

+++ ping360_sonar/include/ping360_sonar/ping360_ros.cpp
@@ -336,7 +336,7 @@ void Ping360ROS::publishPointCloud2(const ros::Time &now){
         *iterY = x[i] * std::sin(sonar_->currentAngle());
         *iterZ = 0;
 
-        *iterIntensity = static_cast<uchar>(data[i]);
+        *iterIntensity = data[i];
 
         // // Increment the iterators
         ++iterX;
